# 2swagger_collection.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CSC309PP1-EndpointsTEST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "noauthAuth": {
        "type": "http",
        "scheme": "noauth"
      }
    }
  },
...
  "paths": {
...
    "/api/hotel/visitor": {
      "get": {
        "tags": ["hotel > visitor"],
        "summary": "search-hotel-and-available",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "starRating",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "5"
          },
          {
            "name": "priceRangeStart",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": "100.00"
          },
          {
            "name": "priceRangeEnd",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "example": "300.00"
          },
          {
            "name": "checkInDate",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2023-10-01"
          },
          {
            "name": "checkOutDate",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "2023-10-10"
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "New York"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "Holiday Inn"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/api/notification": {
      "get": {
        "tags": ["notification"],
        "summary": "Get-notification-id",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "1"
          },
          {
            "name": "messageType",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "HotelBookingRecord"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "delete": {
        "tags": ["notification"],
        "summary": "Delete-notification-id",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  }
}